pyproject.toml
[project]
name = "rfsim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dearpygui-ext>=2.0.0",
    "dearpygui>=2.0.0",
    "networkx>=3.4.2",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "sympy>=1.13.3",
    "cerberus>=1.3.7",
    "pint>=0.24.4",
    "matplotlib>=3.10.1",
    "ruamel-yaml>=0.18.10",
]

[dependency-groups]
dev = [
    "pytest-benchmark>=5.1.0",
    "pytest>=8.3.5",
]


[tool.pytest.ini_options]
pythonpath = ["."]
